Use shared pool across actix-web workers with app_data() approach (#24)
When you run this example on Ryzen 3900 with 12 cores(24 threads) you will get number of connection reached
The situation is that HttServer will start by default some predefined by operating system

https://docs.rs/actix-web/2.0.0/actix_web/struct.HttpServer.html#method.workers

in my case I've got 15
At the same time default pool size is 10

https://docs.rs/r2d2/0.8.8/r2d2/struct.Builder.html#method.max_size

So on my system I needed 150 connections, but default limit in docker-compose.yml is 100.
But based on
https://wiki.postgresql.org/wiki/Number_Of_Database_Connections

A formula which has held up pretty well across a lot of benchmarks for years is that for optimal throughput the number of active connections should be somewhere near ((core_count * 2) + effective_spindle_count). Core count should not include HT threads, even if hyperthreading is enabled

so it just did not make sense to increase number of connections and I did not want to manipulate number of workers and max_size(). So I tried to use app_data() to pass pool as shared across workers and number of connections defaulted to 10. I have also tested using bombardier with 10000 login requests and did not see any degradations of performance.